compleate update and inserting mrc

+++ Models/Mrc.cs
@@ -14,7 +14,7 @@ namespace FarmacyControl.Models
         public Mrc(string nnt, string price)
         {
             this.Nnt = Int32.Parse(nnt);
-            this.Price = Int32.Parse(price);
+            this.Price = (int)Double.Parse(price);
         }
         public Mrc()
         {
